requirements.txt
Django==5.0.11
Pillow==11.0.0
requests==2.31.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
django-shortuuidfield==0.1.3
shortuuid==1.0.13
django-jazzmin==3.0.1
django-rest-knox==5.0.1
gunicorn==20.1.0
django-environ==0.12.0
icecream==2.1.3
django-ckeditor==6.7.0
django-js-asset==3.0.1
supabase==1.2.0
cython==3.0.11
psycopg2-binary==2.9.9
locust==2.17.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-axes==7.0.0
django-ipware==7.0.1
python-ipware==3.0.0
django-location-field== 2.7.3